exercises 4.15-4.23

+++ controller/blogController.js
@@ -1,24 +1,31 @@
 export class BlogController {
-  constructor(Blog) {
+  constructor(Blog, User) {
     this.BlogModel = Blog;
+    this.UserModel = User;
   }
   getAllBlogs = async (req, res, next) => {
     try {
-      const blogs = await this.BlogModel.find({});
+      const blogs = await this.BlogModel.find({}).populate('user', {username: 1});
       res.status(200).json(blogs);
     } catch (e) {
       console.error("Error getting " + e.message);
     }
   };
 
-  createBlog = async (req, res) => {
+  createBlog = async (req, res, next) => {
     const newBlog = { likes: 0, ...req.body };
     try {
-      const blog = new this.BlogModel(newBlog);
+      const user = await this.UserModel.findOne({username: req.session.user.username})
+      const blog = new this.BlogModel({
+        ...newBlog,
+        user: req.session.user.id
+      });
       const createdBlog = await blog.save();
+      user.notes = user.notes.concat(createdBlog._id)
+      await user.save()
       res.status(201).json(createdBlog);
     } catch (e) {
-      res.status(400).json();
+      next(e)
     }
   };
   updateBlog = async (req, res) => {
@@ -37,10 +44,13 @@ export class BlogController {
     const {id} = req.params
 
     try{
+      const {user} = await this.BlogModel.findById(id)
+     
+      if( user.toString() !== req.session.user.id.toString() ) return res.status(401).json({message: "you are not the author"})
       await this.BlogModel.findByIdAndDelete(id)
       res.status(204).json({})
     }catch(e){
-      res.status(404).json({})
+      console.error(e.message)
     } 
   }
 }
